requirements.txt
fastapi==0.85.0
uvicorn[standard]==0.15.0
PyJWT==2.5.0
Werkzeug==2.2.2
pandas==1.5.0

requests==2.28.1
asgiref==3.5.2
pyyaml==6.0

cryptography==38.0.1

untangle>=1.2.1
ultralytics>=8.0.200